Adjust architecture to avoid mutual dependency b/w data model and metric abstraction
